[Refactor] HikariCP 구조 변경
