+++ app/src/main/java/kr/co/lion/modigm/db/study/StudyDao.kt
@@ -1,66 +1,20 @@
 package kr.co.lion.modigm.db.study
 
 import android.util.Log
-import com.zaxxer.hikari.HikariConfig
-import com.zaxxer.hikari.HikariDataSource
-import kotlinx.coroutines.CoroutineExceptionHandler
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.async
-import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.withContext
-import kr.co.lion.modigm.BuildConfig
+import kr.co.lion.modigm.db.HikariCPDataSource
 import kr.co.lion.modigm.model.SqlStudyData
-import java.sql.Connection
 
 class StudyDao {
 
     private val tag = "StudyDao"
 
-    // HikariCP 설정을 초기화하는 suspend 함수
-    private suspend fun initDataSource(): HikariDataSource = withContext(Dispatchers.IO) {
-        val hikariConfig = HikariConfig().apply {
-            jdbcUrl = BuildConfig.DB_URL
-            username = BuildConfig.DB_USER
-            password = BuildConfig.DB_PASSWORD
-            driverClassName = "com.mysql.jdbc.Driver"
-            maximumPoolSize = 10
-            minimumIdle = 10
-            connectionTimeout = 30000 // 30초
-            idleTimeout = 600000 // 10분
-            maxLifetime = 1800000 // 30분
-            validationTimeout = 5000 // 5초
-            leakDetectionThreshold = 30000 // 30초
-        }
-        HikariDataSource(hikariConfig)
-    }
-
-    private val coroutineExceptionHandler = CoroutineExceptionHandler { _, throwable ->
-        Log.e(tag, "HikariCP coroutineExceptionHandler 에러 ", throwable)
-    }
-
-    private val job = Job()
-    private val coroutineScope: CoroutineScope = CoroutineScope(Dispatchers.IO + job + coroutineExceptionHandler)
-    private val dataSourceDeferred: Deferred<HikariDataSource> by lazy {
-        coroutineScope.async {
-            initDataSource()
-        }
-    }
-
-    private suspend fun getConnection(): Connection {
-        val dataSource = dataSourceDeferred.await()
-        return withContext(Dispatchers.IO) {
-            dataSource.connection
-        }
-    }
-
     // 스터디와 스터디 멤버 데이터 조회 (좋아요 여부 포함)
     suspend fun selectAllStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> =
         withContext(Dispatchers.IO) {
             runCatching {
-                getConnection().use { connection ->
+                HikariCPDataSource.getConnection().use { connection ->
                     val combinedQuery = """
                 SELECT s.*, COUNT(sm.userIdx) as memberCount, 
                        IF(f.favoriteIdx IS NOT NULL, TRUE, FALSE) as isFavorite
@@ -94,7 +48,7 @@ class StudyDao {
     suspend fun selectMyStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> =
         withContext(Dispatchers.IO) {
             runCatching {
-                getConnection().use { connection ->
+                HikariCPDataSource.getConnection().use { connection ->
                     val query = """
                     SELECT s.*, COUNT(sm.userIdx) as memberCount,
                            IF(f.favoriteIdx IS NOT NULL, TRUE, FALSE) as isFavorite
@@ -128,7 +82,7 @@ class StudyDao {
     suspend fun selectFavoriteStudyData(userIdx: Int): Result<List<Triple<SqlStudyData, Int, Boolean>>> =
         withContext(Dispatchers.IO) {
             runCatching {
-                getConnection().use { connection ->
+                HikariCPDataSource.getConnection().use { connection ->
                     val query = """
                 SELECT s.*, COUNT(sm.userIdx) as memberCount, 
                        IF(f.favoriteIdx IS NOT NULL, TRUE, FALSE) as isFavorite
@@ -161,7 +115,7 @@ class StudyDao {
     suspend fun addFavorite(userIdx: Int, studyIdx: Int): Result<Boolean> =
         withContext(Dispatchers.IO) {
             runCatching {
-                getConnection().use { connection ->
+                HikariCPDataSource.getConnection().use { connection ->
                     val queryInsert = "INSERT INTO tb_favorite (userIdx, studyIdx) VALUES (?, ?)"
                     connection.prepareStatement(queryInsert).use { statementInsert ->
                         statementInsert.setInt(1, userIdx)
@@ -180,7 +134,7 @@ class StudyDao {
     suspend fun removeFavorite(userIdx: Int, studyIdx: Int): Result<Boolean> =
         withContext(Dispatchers.IO) {
             runCatching {
-                getConnection().use { connection ->
+                HikariCPDataSource.getConnection().use { connection ->
                     val queryDelete = "DELETE FROM tb_favorite WHERE userIdx = ? AND studyIdx = ?"
                     connection.prepareStatement(queryDelete).use { statementDelete ->
                         statementDelete.setInt(1, userIdx)
@@ -194,19 +148,4 @@ class StudyDao {
                 Result.failure<Boolean>(e)
             }
         }
-
-    suspend fun close() {
-        try {
-            // Job을 취소하고 모든 하위 코루틴이 종료될 때까지 대기
-            job.cancelAndJoin()
-            // 데이터 소스가 완료된 경우, 데이터 소스를 안전하게 종료
-            withContext(Dispatchers.IO) {
-                if (dataSourceDeferred.isCompleted) {
-                    dataSourceDeferred.await().close()
-                }
-            }
-        } catch (e: Exception) {
-            Log.e(tag, "HikariCP 코루틴 취소 실패", e)
-        }
-    }
 }
